+++ checkout-pricing/src/main/java/com/tdd/api/rest/evaluate/EvaluateRequest.java
@@ -3,6 +3,6 @@ package com.tdd.api.rest.evaluate;
 import java.util.Map;
 
 public class EvaluateRequest {
-    public Map<String, Long> cart;
+    public Map<String, Integer> cart;
     public String ruleSet;
 }
